testpla/requirements.txt
appdirs==1.4.3
argcomplete==1.10.0
asgiref==3.2.3
attrs==19.3.0
beautifulsoup4==4.8.0
black==19.10b0
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==3.0.2
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
gunicorn==19.9.0
IMAPClient==2.1.0
lxml==4.4.2
nltk==3.4.5
olefile==0.46
pathspec==0.7.0
pdfminer.six==20181108
Pillow==6.2.1
psycopg2==2.8.2
pycryptodome==3.9.4
PyPDF2==1.26.0
python-dateutil==2.8.0
python-pptx==0.6.18
pytz==2019.3
regex==2019.12.20
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
sqlparse==0.3.0
textract==1.6.3
toml==0.10.0
topia.termextract==1.1.0
typed-ast==1.4.0
tzlocal==1.5.1
whitenoise==4.1.2
xlrd==1.2.0
XlsxWriter==1.2.7
zope.interface==4.7.1
